HV-1155 Move SetContextClassLoader privileged action to a shared package
